+++ modules/dggs/src/s2.ts
@@ -1,0 +1,6 @@
+// math.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export {S2Decoder} from './s2-decoder';
+export {getS2ChildIndex} from './s2-geometry/s2-token';